refactor: carRacingView 네이밍 수정
